fix: use canvas renderer for hero animation
