+++ package.json
@@ -32,7 +32,6 @@
     "@docusaurus/plugin-sitemap": "2.4.1",
     "@docusaurus/preset-classic": "2.4.1",
     "@docusaurus/theme-mermaid": "2.4.1",
-    "@dotlottie/player-component": "1.1.1",
     "@fec/remark-a11y-emoji": "3.1.0",
     "@mdx-js/react": "1.6.22",
     "@octokit/core": "4.2.1",
@@ -48,6 +47,7 @@
     "prism-react-renderer": "2.0.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-lottie-player": "1.5.4",
     "react-modal": "3.16.1",
     "react-netlify-forms": "1.3.3",
     "react-wrap-balancer": "0.5.0",
