+++ src/components/ModulePage/Header/index.tsx
@@ -1,6 +1,6 @@
 import { PageMetadata } from '@docusaurus/theme-common';
 import baseStyles from '@site/src/pages/styles.module.css';
-import { getImage } from '@site/src/utils/socialImageHelper';
+import { getImage } from '@site/src/utils/socialImageUtils';
 import clsx from 'clsx';
 import React from 'react';
 import Balancer from 'react-wrap-balancer';
