+++ src/pages/datalink/index.tsx
@@ -95,8 +95,10 @@ export default function DatalinkPage(): JSX.Element {
               </Balancer>
             </p>
           </ModulePageCTA>
-          <ContactForm />
         </main>
+        <div className={baseStyles.contentBottom}>
+          <ContactForm />
+        </div>
       </Layout>
     </>
   );
