+++ src/pages/index.tsx
@@ -6,10 +6,11 @@ import AwsLogo from '@site/src/img/providers/aws.svg';
 import DigitalOceanLogo from '@site/src/img/providers/digitalocean.svg';
 import GoogleCloudLogo from '@site/src/img/providers/google-cloud.svg';
 import KubernetesLogo from '@site/src/img/providers/kubernetes.svg';
-import heroLottie from '@site/src/lottie/hero.lottie';
+import heroLottie from '@site/src/lottie/hero.json';
 import Layout from '@theme/Layout';
 import { clsx } from 'clsx';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
+import Lottie from 'react-lottie-player';
 import Balancer from 'react-wrap-balancer';
 import styles from './styles.module.css';
 
@@ -19,9 +20,6 @@ export default function Home(): JSX.Element {
   const [isMobile, setIsMobile] = useState(true);
   // const [prefersReducedMotion, setPrefersReducedMotion] = useState(false);
 
-  const onAnimationReady = useCallback(() => {
-    heroAnimationRef.current?.getLottie()?.setSubframe(false);
-  }, []);
   const onResize = useCallback(
     () => setIsMobile(window.innerWidth <= mobileMaxWidth),
     []
@@ -29,8 +27,6 @@ export default function Home(): JSX.Element {
   // const onReducedMotion = useCallback(() => setPrefersReducedMotion(true), []);
 
   useEffect(() => {
-    require('@dotlottie/player-component');
-
     setIsMobile(window.innerWidth <= mobileMaxWidth);
     window.addEventListener('resize', onResize);
 
@@ -43,23 +39,6 @@ export default function Home(): JSX.Element {
       // mediaQuery.removeEventListener('change', onReducedMotion);
     };
   }, [onResize]);
-
-  useEffect(() => {
-    let ref = null;
-
-    if (!isMobile && heroAnimationRef.current) {
-      ref = heroAnimationRef.current;
-      ref.addEventListener('ready', onAnimationReady);
-      ref.load(heroLottie, {
-        progressiveLoad: true,
-        runExpressions: false,
-      });
-    }
-
-    return () => {
-      ref?.removeEventListener('ready', onAnimationReady);
-    };
-  }, [isMobile, onAnimationReady]);
 
   return (
     <>
@@ -115,21 +94,29 @@ export default function Home(): JSX.Element {
             </div>
             {!isMobile ? (
               <div className={styles.heroRight}>
-                <div className={styles.heroAnim} aria-hidden="true">
-                  <div>
-                    <dotlottie-player
-                      ref={heroAnimationRef}
-                      autoplay
-                      loop
-                      background="#fff"
-                    />
-                  </div>
-                </div>
+                <Lottie
+                  animationData={heroLottie}
+                  ref={heroAnimationRef}
+                  useSubframes={false}
+                  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                  /* @ts-ignore */
+                  renderer="canvas"
+                  rendererSettings={{
+                    preserveAspectRatio: 'xMidYMid slice',
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                    /* @ts-ignore */
+                    runExpressions: false,
+                  }}
+                  play
+                  loop
+                  className={styles.heroAnim}
+                  aria-hidden="true"
+                />
               </div>
             ) : null}
           </div>
         </header>
-        <main className={styles.homeContent}>
+        <main className={styles.contentBottom}>
           <section className={styles.section}>
             <div className={styles.inner}>
               <h2>
