+++ src/pages/metricmate/index.tsx
@@ -95,8 +95,10 @@ export default function MetricmatePage(): JSX.Element {
               </Balancer>
             </p>
           </ModulePageCTA>
-          <ContactForm />
         </main>
+        <div className={baseStyles.contentBottom}>
+          <ContactForm />
+        </div>
       </Layout>
     </>
   );
