+++ src/pages/rewind/index.tsx
@@ -66,8 +66,10 @@ export default function RewindPage(): JSX.Element {
               </Balancer>
             </p>
           </ModulePageCTA>
-          <ContactForm />
         </main>
+        <div className={baseStyles.contentBottom}>
+          <ContactForm />
+        </div>
       </Layout>
     </>
   );
