+++ src/pages/spentwise/index.tsx
@@ -83,8 +83,10 @@ export default function SpentwisePage(): JSX.Element {
               </Balancer>
             </p>
           </ModulePageCTA>
-          <ContactForm />
         </main>
+        <div className={baseStyles.contentBottom}>
+          <ContactForm />
+        </div>
       </Layout>
     </>
   );
