+++ src/pages/styles.module.css
@@ -100,22 +100,8 @@ html[data-theme='dark'] .hero {
 
 .heroAnim {
   position: absolute;
-  left: -0.5%;
-  border-radius: 50%;
   width: 100%;
   height: 100%;
-  overflow: hidden;
-}
-
-.heroAnim > div {
-  position: absolute;
-  top: 0.15%;
-  margin: -0.5%;
-}
-
-.heroAnim dotlottie-player {
-  border-radius: 50%;
-  overflow: hidden;
 }
 
 .buttons {
@@ -218,7 +204,7 @@ html[data-theme='dark'] .docsButton::before {
   background-image: url('./img/icons/docs-light.svg');
 }
 
-.homeContent {
+.contentBottom {
   background-image: url('./img/bg-light.webp');
   background-position: top;
   background-size: cover;
@@ -226,7 +212,7 @@ html[data-theme='dark'] .docsButton::before {
   padding: 24px 0 5rem;
 }
 
-html[data-theme='dark'] .homeContent {
+html[data-theme='dark'] .contentBottom {
   background-image: url('./img/bg-dark.webp');
 }
