+++ src/theme/BlogListPage.tsx
@@ -4,7 +4,7 @@ import {
   ThemeClassNames,
 } from '@docusaurus/theme-common';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import { getImage } from '@site/src/utils/socialImageHelper';
+import { getImage } from '@site/src/utils/socialImageUtils';
 import BlogLayout from '@theme/BlogLayout';
 import type { Props } from '@theme/BlogListPage';
 import BlogListPaginator from '@theme/BlogListPaginator';
