+++ src/theme/DocItem/Metadata.tsx
@@ -1,6 +1,6 @@
 import { PageMetadata } from '@docusaurus/theme-common';
 import { useDoc } from '@docusaurus/theme-common/internal';
-import { getImage } from '@site/src/utils/socialImageHelper';
+import { getImage } from '@site/src/utils/socialImageUtils';
 import React, { useEffect } from 'react';
 
 export default function DocItemMetadata(): JSX.Element {
